Added a hasher template for valid key value types. Implemented hashing for node creation and insertion for key hash index. Implemented more tests for Node: string_char key/value types, and hashing index tests

+++ src/Node/Node.h
@@ -19,7 +19,7 @@ struct Record
 {
     K record_key;
     V record_value;
-    float record_index;
+    unsigned int record_index;
     float timeStamp;
 };
